+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fix support for non-scalar keys in YAML mappings (eg: "1: car").
+
 ## 1.3.0 - 2018-07-30
 ### Added
 - Add a `RawSource` option that selectively disables variable expansion.